requirements.txt
Django==2.2.2
gunicorn==19.9.0
djangorestframework==3.10.2
psycopg2=2.8.3
